refactor Timer component
