+++ components/Timer/Timer.jsx
@@ -1,5 +1,8 @@
 import { useState, useEffect } from "react";
-import { View, Text, StyleSheet } from "react-native";
+import { View, Text} from "react-native";
+import styles from "./TimerStyles";
+import { formatTime } from "./algorithms";
+
 
 const Timer = ({ timeRemaining, isWhiteTurn, isWhite }) => {
   const [seconds, setSeconds] = useState(timeRemaining);
@@ -16,11 +19,6 @@ const Timer = ({ timeRemaining, isWhiteTurn, isWhite }) => {
     };
   }, [isWhiteTurn]);
 
-  const formatTime = (seconds) => {
-    let minutes = Math.floor(seconds / 60);
-    let sec = seconds % 60;
-    return `${minutes < 10 ? "0" : ""}${minutes}:${sec < 10 ? "0" : ""}${sec}`;
-  };
   return (
     <View style={styles.timerContainer}>
       <Text style={styles.timerText}>{formatTime(seconds)}</Text>
@@ -28,18 +26,5 @@ const Timer = ({ timeRemaining, isWhiteTurn, isWhite }) => {
   );
 };
 
-const styles = StyleSheet.create({
-  timerContainer: {
-    height: "50%",
-    backgroundColor: "yellow",
-    padding: 10,
-    margin: 10,
-    flexDirection: "row",
-    alignItems: "center"
-  },
-  timerText: {
-    color: "black"
-  }
-});
 
 export default Timer;
