+++ components/Timer/TimerStyles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    timerContainer: {
+      height: "50%",
+      backgroundColor: "yellow",
+      padding: 10,
+      margin: 10,
+      flexDirection: "row",
+      alignItems: "center"
+    },
+    timerText: {
+      color: "black"
+    }
+  });
+
+  export default styles;
+
